pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmbench"
version = "0.1.0"
description = "LLM API Benchmark Tool in Python"
license = "MIT"
authors = [
    {name = "Dario cargnino", email = "dario.cargnino@gmail.com"},
]
dependencies = [
    "openai>=1.0.0",
    "typer>=0.9.0",
    "PyYAML>=6.0.0",
    "tqdm>=4.65.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "rich>=13.7.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[project.scripts]
llmbench = "cli.main:app"